requirements.txt
PyQt5
opencv-python
numpy
pillow
scipy
PyQt5-stubs
pyaudio
google-generativeai
python-dotenv
